address some compile warnings

+++ src/localgraph.cpp
@@ -323,7 +323,7 @@ std::vector<LocalNodePtr> LocalGraph::nodes_along_string(const std::string &quer
             }
             if (strcasecmp(query_string.c_str(), candidate_string.c_str()) == 0) {
                 return p;
-            } else if (candidate_string.size() > longest_length) {
+            } else if (candidate_string.size() > (uint) longest_length) {
                 longest_path = p;
                 longest_length = candidate_string.size();
             }
